requirements.txt
langgraph==0.0.44
langchain==0.1.10
google-cloud-bigquery==3.17.2
google-generativeai==0.3.2
python-dotenv==1.0.0
pytest==8.0.0
regex==2023.12.25